Add prereq check to pem_create_keys.sh

+++ pem_create_keys.sh
@@ -2,9 +2,40 @@
 
 # Script to create the test keys used pem_test.cxx
 
+##################################
+# prerequisites
+
+if [[ -z "$CXX" ]]; then
+    CXX=g++
+fi
+
+if [[ -z $(command -v "$CXX") ]]; then
+    echo "Please install a compiler like g++"
+    exit 1
+fi
+
+if [[ -z $(command -v openssl) ]]; then
+    echo "Please install openssl package"
+    exit 1
+fi
+
+if [[ -z $(command -v perl) ]]; then
+    echo "Please install perl package"
+    exit 1
+fi
+
+##################################
+# test program
+
 echo "Compiling test program"
+
 rm -rf pem_test.exe
-g++ pem_test.cxx ./libcryptopp.a -o pem_test.exe
+"$CXX" pem_test.cxx ./libcryptopp.a -o pem_test.exe
+
+##################################
+# Test keys
+
+echo "Generating OpenSSL keys"
 
 # RSA private key, public key, and encrypted private key
 openssl genrsa -out rsa-priv.pem 1024
@@ -25,12 +56,11 @@ openssl ec -in ec-priv.pem -out ec-enc-priv.pem -aes128 -passout pass:test
 
 openssl dhparam -out dh-params.pem 512
 
+##################################
+# malformed
+
 # Only the '-----BEGIN PUBLIC KEY-----'
 echo "-----BEGIN PUBLIC KEY-----" > rsa-short.pem
-
-# Two keys in one file
-cat rsa-pub.pem > rsa-pub-double.pem
-cat rsa-pub.pem >> rsa-pub-double.pem
 
 # Removes last CR or LF (or CRLF)
 perl -pe 'chomp if eof' rsa-pub.pem > rsa-trunc-1.pem
@@ -42,8 +72,10 @@ perl -pe 'chop if eof' rsa-trunc-1.pem > rsa-trunc-2.pem
 cat rsa-trunc-1.pem > rsa-concat.pem
 cat rsa-pub.pem >> rsa-concat.pem
 
-# Download cacert.pem. Its got 150 or so we can parse
+##################################
+# cacert.pem
 
 if [ ! -e "cacert.pem" ]; then
-  wget http://curl.haxx.se/ca/cacert.pem -O cacert.pem
+    wget http://curl.haxx.se/ca/cacert.pem -O cacert.pem
 fi
+
